% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictModel.R
\name{predictModel}
\alias{predictModel}
\title{function for getting modeled incidence and observations used for training}
\usage{
predictModel(query)
}
\arguments{
\item{query}{JSON string with description of output data requested from model}
}
\value{
A tibble with requested data, formatted to translate nicely to hierarchical json
}
\description{
This function loads a cached model object containing two datasets
  (1) observed incidence of flu in ILI
  (2) modeled incidence after statistical smoothing
}
\examples{
query <- jsonlite::toJSON(  list(output='incidence', attributes=c('Census_Tract','vax_status'), basis=c('year'), values=c('incidence_median','incidence_sd') ))
predictModel(query)

}
